AI/requirements.txt
fastapi==0.115.12
openpyxl==3.1.5
pydantic==2.11.5
python-dotenv==1.1.0
typing_extensions==4.14.0
openai 
gunicorn
uvicorn
